// src/pdf-service/dto/getData.dto.ts
/* eslint-disable @typescript-eslint/no-unused-vars */
export class getDataDto {
  id: string;
  mb_idtrans: string;
  namaTpg: string;
  fullname: string;
  mutu: string;
  jumlah: string;
  harga_dasar: string;
  harga_tambahan: string;
  total: string;
  namaPenyadap: string;
  created_at: string;
}